:recycle: refactor: base response 위치 변경
base response 는 exception 을 위해 존재하는 dto가 아니기 때문에 따로 폴더로 분리했습니다.
또한 페이지네이션 dto도 추가할 예정이어서 폴더로 따로 분리했습니다.
