+++ src/main/java/com/bookbook/booklink/auth_service/controller/TokenController.java
@@ -1,6 +1,6 @@
 package com.bookbook.booklink.auth_service.controller;
 
-import com.bookbook.booklink.common.exception.BaseResponse;
+import com.bookbook.booklink.common.dto.BaseResponse;
 import com.bookbook.booklink.common.exception.CustomException;
 import com.bookbook.booklink.common.exception.ErrorCode;
 import com.bookbook.booklink.common.jwt.model.RefreshToken;
