+++ src/main/java/com/bookbook/booklink/library_service/controller/docs/LibraryApiDocs.java
@@ -1,7 +1,7 @@
 package com.bookbook.booklink.library_service.controller.docs;
 
 import com.bookbook.booklink.common.exception.ApiErrorResponses;
-import com.bookbook.booklink.common.exception.BaseResponse;
+import com.bookbook.booklink.common.dto.BaseResponse;
 import com.bookbook.booklink.common.exception.ErrorCode;
 import com.bookbook.booklink.library_service.model.dto.request.LibraryRegDto;
 import com.bookbook.booklink.library_service.model.dto.request.LibraryUpdateDto;
