+++ src/main/java/com/bookbook/booklink/notification_service/controller/docs/NotificationApiDocs.java
@@ -1,7 +1,7 @@
 package com.bookbook.booklink.notification_service.controller.docs;
 
 import com.bookbook.booklink.common.exception.ApiErrorResponses;
-import com.bookbook.booklink.common.exception.BaseResponse;
+import com.bookbook.booklink.common.dto.BaseResponse;
 import com.bookbook.booklink.common.exception.ErrorCode;
 import com.bookbook.booklink.notification_service.model.dto.response.NotificationResDto;
 import io.swagger.v3.oas.annotations.Operation;
