+++ src/main/java/com/bookbook/booklink/point_service/controller/docs/PointApiDocs.java
@@ -1,7 +1,7 @@
 package com.bookbook.booklink.point_service.controller.docs;
 
 import com.bookbook.booklink.common.exception.ApiErrorResponses;
-import com.bookbook.booklink.common.exception.BaseResponse;
+import com.bookbook.booklink.common.dto.BaseResponse;
 import com.bookbook.booklink.common.exception.ErrorCode;
 import com.bookbook.booklink.point_service.model.dto.request.PointUseDto;
 import com.bookbook.booklink.point_service.model.dto.response.PointBalanceDto;
